+++ CHANGELOG.md
@@ -19,6 +19,8 @@ therefore are not "truly breaking".
   - `node_neighbors` -> `nearby_positions`
   - `get_node_contents` -> `agents_in_pos`
   - `pick_empty` -> `random_empty`
+  - `find_empty_nodes` -> `empty_positions`
+  - `has_empty_nodes` -> `has_empty_positions`
 
 # v3.7
 - Add the ability to decide whether the agent step or the model step should be performed first using the `agents_first` argument.